storage adapter abstract class
