+++ experimental/emulated_vault/opt/emulated-vault/storage/fs_adapter.py
@@ -16,9 +16,9 @@
 #
 
 import os
-from . base import Base
+from . adapter_base import AdapterBase
 
-class Fs(Base):
+class FsAdapter(AdapterBase):
 	"""
 	Fs (file system) adapter class.
 	This class implements the API required for storing and retriving the content kept in the vault.
@@ -26,43 +26,20 @@ class Fs(Base):
 	Inputs configuration (held under "fs-adapter" section in the config file) include:
 	:param db-base-os-path: The path in which the DB files are stored
 	:param ping-os-path: The path of a variable mimicing the RIAK ping functionality
-
-	Interface methods 
-	:meth:`_get_parameter_storage_path` given a url-key of a parameter return the storage path
-	:meth:`_get_parameter_storage_path_from_url_path` given a storage path of a parameter
-	retrun the url-key
-	:meth:`_init_cfg` given a config-parser object, read the parameters required for 
-	adapter's operation. Return "success" boolean value
-	:meth:`_init_ping` prepare the adapter for beign able to provide info to reply for 
-	"ping" requests. Return "success" boolean value
-	:meth:`_ping` test the 'ping' status with the adapter. 
-	Return a tuple: "success" boolean & "value" kept as ping variable
-	:meth:`_read_parameter_by_storage_path` given a storage path retrieve the parameter value.
-	Return a tuple: "success" boolean & "value" kept in the parameter
-	:meth:`_read_parameters_by_storage_path` given a storage and and a key holding 
-	filters on the key and values.
-	Return "success" boolean indicating a sucessful write, and a key->value dictionary 
-	for the relevant parameters
-	:meth:`_write_parameter_by_storage_path` given a storage path and a value string,
-	keep the parameter value. Return "success" boolean indicating a sucessful write
-	:meth:`_remove_parameter_by_storage_path` given a storage path, delete the parameter
-	from the DB. Return "success" boolean indicating a sucessful deletion		
 	"""
+	
 	def __init__ (self, logger):
 		"""
 		The class constructor.
 		:param logger: logger to send log messages to
 		:type logger: a python logging logger class
 		"""
-		Base.__init__(self, logger)
+		self.logger = logger
 
-	def _init_cfg(self, fullConfig):
+	def init_cfg(self, fullConfig):# -> bool:
 		"""
 		Initialize the class basic parameters. Part of Adapter required API.
-		:param fullConfig: configuration to operate upon.
-		:type fullConfig: configparser.ConfigParser class
-		:return: 'True' for successful initialization
-		:rtype: Boolean
+		Read the relevant section in the configuration
 		"""
 		myCfgData = dict(fullConfig.items("fs-adapter")) if fullConfig.has_section("fs-adapter") else {}
 		self.basePath = myCfgData.get("db-base-os-path")
@@ -75,59 +52,43 @@ class Fs(Base):
 			return False 
 		return True
 
-	def _get_parameter_storage_path(self, parameterUrlPath):
-		"""
-		Conversion function - taking a key's path and translate to a file path on the file system
-		:param parameterUrlPath: the "url-path" like key of the variable
-		:type parameterUrlPath: str
-		:return: file path of where the value is be kept
-		:rtype: str
-		"""
-		return os.path.join(self.basePath, parameterUrlPath.lstrip("/").replace("/", os.path.sep))
-
-	def _get_parameter_storage_path_from_url_path(self, parameterStoragePath):
-		"""
-		Conversion function - taking file path on the file system and translate to key's path
-		:param parameterStoragePath: the file name holding a value
-		:type parameterUrlPath: str
-		:return: the matching variable url-path like key
-		:rtype: str
-		"""
-		return "/"+os.path.relpath(parameterStoragePath, self.basePath).replace(os.path.sep, "/")
-
-	def _init_ping(self):
+	def init(self):# -> bool:
 		"""
 		Initialize the class ability to answer for ping. Part of Adapter required API.
-		:return: 'True' for successful initialization
-		:rtype: Boolean
 		"""
-		value = "OK"
-		success = self._write_parameter_by_storage_path(self.pingStoragePath, value)
+		value = ":)"
+		success = self.write_parameter_by_storage_path(self.pingStoragePath, value)
 		if not success:
 			self.logger.error("Failed to set parameter %s", self.pingStoragePath)
 			return False
 		return True
 
-	def _ping(self):
+	def get_parameter_storage_path(self, parameterUrlPath):# -> str:
 		"""
-		get value for the ping request. Part of Adapter required API.
-		:return: A tuple - 'True' for successful retrival and the retrieved value
-		:rtype: Tuple[Boolean, str]
+		Conversion function - taking a key's path and translate to a file path on the file system
 		"""
-		success, value = self._read_parameter_by_storage_path(self.pingStoragePath)
+		return os.path.join(self.basePath, parameterUrlPath.lstrip("/").replace("/", os.path.sep))
+
+	def get_parameter_storage_path_from_url_path(self, parameterStoragePath):# -> str:
+		"""
+		Conversion function - taking file path on the file system and translate to key's path
+		"""
+		return "/"+os.path.relpath(parameterStoragePath, self.basePath).replace(os.path.sep, "/")
+
+	def ping(self):# -> bool:
+		"""
+		Get value for the ping request by its path. Part of Adapter required API.
+		"""
+		success, value = self.read_parameter_by_storage_path(self.pingStoragePath)
 		if not success or value is None:
 			self.logger.error("no ping response")
-			return (False, "")
+			return False
 		self.logger.debug("ping response: %s", value)
-		return (True, value)
+		return True
 
-	def _read_parameter_by_storage_path(self, parameterStoragePath):
+	def read_parameter_by_storage_path(self, parameterStoragePath):# -> (bool, str):
 		"""
 		Reading the value from the provided file name.
-		:param parameterStoragePath: the file name 
-		:type parameterStoragePath: str
-		:return: 'True' for successful retrivaland the retrieved value
-		:rtype: Tuple[boolean, str]
 		"""
 		self.logger.debug("Get parameter by os path: %s", parameterStoragePath)
 		try:
@@ -140,16 +101,9 @@ class Fs(Base):
 		self.logger.debug("Get parameter by os path %s succeed", parameterStoragePath)
 		return True, value
 
-	def _read_parameters_by_storage_path(self, parameterStoragePathPrefix, keyFilters):
+	def read_parameters_by_storage_path(self, parameterStoragePathPrefix, keyFilters):# -> (bool, dict(str, str)):
 		"""
 		Reading the values of the parameters the provided directory.
-		:param parameterStoragePathPrefix: the directory to look into
-		:type parameterStoragePathPrefix: str
-		:param keyFilters: filter-name/filter-func dict, holding functions that get a key as 
-		input and retunn "true" if key should be included in the result
-		:type keyFilters: Dict[str,function[str]]
-		:return: 'True' for successful retrival and a dict for key-name/value 
-		:rtype: Tuple[boolean, Dict[str, str]]
 		"""
 		self.logger.debug("Get parameters under os path %s", parameterStoragePathPrefix)
 		parameters = {}
@@ -162,14 +116,14 @@ class Fs(Base):
 			filteredOut = False
 			for filterName, filterfunc in keyFilters.items():#items() - supporting python 2&3
 				if not filterfunc(fileName):
-					self.logger.debug("Parameter os path %s dropped, not matching filter %s", self._get_parameter_storage_path_from_url_path(fileName), filterName)
+					self.logger.debug("Parameter os path %s dropped, not matching filter %s", self.get_parameter_storage_path_from_url_path(fileName), filterName)
 					filteredOut = True
 					break
 			if filteredOut:
 				continue
 
-			parameterUrlPath = self._get_parameter_storage_path_from_url_path(fileName)
-			success, value = self._read_parameter_by_storage_path(fileName)
+			parameterUrlPath = self.get_parameter_storage_path_from_url_path(fileName)
+			success, value = self.read_parameter_by_storage_path(fileName)
 			if not success:
 				self.logger.error("%s parameter os path not found.", fileName)
 				return False, None
@@ -179,15 +133,9 @@ class Fs(Base):
 		return True, parameters
 
 
-	def _write_parameter_by_storage_path(self, parameterStoragePath, value):
+	def write_parameter_by_storage_path(self, parameterStoragePath, value):# -> bool:
 		"""
 		Writing the value to the provided file name.
-		:param parameterStoragePath: the file name 
-		:type parameterStoragePath: str
-		:param value: value to be writen
-		:type value: str
-		:return: 'True' for successful writing
-		:rtype: Boolean		
 		"""
 		self.logger.debug("Set parameter by os path %s", parameterStoragePath)
 		try:
@@ -202,13 +150,9 @@ class Fs(Base):
 		self.logger.debug("Set parameter os path %s done", parameterStoragePath)
 		return True
 
-	def _remove_parameter_by_storage_path(self, parameterStoragePath):
+	def remove_parameter_by_storage_path(self, parameterStoragePath):# -> bool:
 		"""
 		Deleting the the provided file.
-		:param parameterStoragePath: the file name 
-		:type parameterStoragePath: str
-		:return: 'True' for successful deletion
-		:rtype: Boolean		
 		"""
 		self.logger.debug("Delete parameter os path %s", parameterStoragePath)
 		try:
